Fix missing versions in Aboutt Modal
Move AboutModal to be a component

Fix-Issue: AAH-161
